+++ CHANGELOG.md
@@ -25,7 +25,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Make the fields in the FE editor configurable (#4362)
 - Add an `Event.slug` property (#4331)
 - Add `Registration::belongsToUser()` (#4255)
-- Provide downloads in the "my registrations" view (#4249, #4250, #4251)
+- Provide downloads in the "my registrations" view (#4249, #4250, #4251, #4461)
 - Add `Event.downloadsForAttendees` (#4243)
 - Add unregistration to the `my registrations` single view (#4236)
 - Add a reimplemented "my registrations" plugin